decoder: use build flags to export NewDecoderSize only on go >= 1.6
